refactor: Moved UrlUtility to utils/url_utils
